fix: other tiny fixes

+++ src/math/abstracts/Vault.sol
@@ -26,9 +26,9 @@ abstract contract Vault is MaxDepositMintStateToData, MaxWithdrawRedeemStateToDa
         uint256 borrowPrice,
         uint8 borrowTokenDecimals
     ) internal pure returns (uint256) {
-        // casting to uint256 is safe because collateral is considered to be greater than borrow
+        // casting to uint256 is safe because collateral and borrow are considered to be greater than zero
         // forge-lint: disable-next-line(unsafe-typecast)
-        uint256 totalAssetsInUnderlying = uint256(collateral - borrow);
+        uint256 totalAssetsInUnderlying = uint256(collateral) - uint256(borrow);
 
         if (totalAssetsInUnderlying >= maxTotalAssetsInUnderlying) {
             return 0;
